@@ -10,6 +10,7 @@ import {
   getterOnly,
   hasOwnProperty,
   setAsyncState,
+  vueSet
 } from './util'
 import { getWatchedGetter } from './watch'
 import {
@@ -19,13 +20,18 @@ import {
 
 const prefix = '_async_computed$'
 
+/** @type {import('vue').Plugin} */
 const AsyncComputed = {
-  install (Vue, pluginOptions) {
-    Vue.config
-      .optionMergeStrategies
-      .asyncComputed = Vue.config.optionMergeStrategies.computed
+  install (app, pluginOptions) {
+    // Vue 2 exposes the `computed` merge strategy, but Vue 3 does not.
+    // Vue 3 calls `Object.assign` eventually though, so we can use that.
+    // See: https://github.com/vuejs/core/blob/32bdc5d1900ceb8df1e8ee33ea65af7b4da61051/packages/runtime-core/src/componentOptions.ts#L1059
+    const mergeStrategy = app.config.optionMergeStrategies.computed || (function (to, from) {
+      return to ? Object.assign(Object.create(null), to, from) : from
+    })
+    app.config.optionMergeStrategies.asyncComputed = mergeStrategy
 
-    Vue.mixin(getAsyncComputedMixin(pluginOptions))
+    app.mixin(getAsyncComputedMixin(pluginOptions))
   }
 }
 
@@ -92,7 +98,7 @@ function handleAsyncComputedPropetyChanges (vm, key, pluginOptions) {
       if (thisPromise !== promiseId) return
 
       setAsyncState(vm, key, 'error')
-      vm.$set(vm.$data._asyncComputed[key], 'exception', err)
+      vueSet(vm, vm.$data._asyncComputed[key], 'exception', err)
       if (pluginOptions.errorHandler === false) return
 
       const handler = (pluginOptions.errorHandler === undefined)
@@ -106,10 +112,17 @@ function handleAsyncComputedPropetyChanges (vm, key, pluginOptions) {
       }
     })
   }
-  vm.$set(vm.$data._asyncComputed, key, {
+  vueSet(vm, vm.$data._asyncComputed, key, {
     exception: null,
     update: () => {
-      if (!vm._isDestroyed) {
+      const mounted = (
+        // Vue 2
+        vm._isDestroyed === false ||
+        // Vue 3
+        // TODO: I think this might be dev environment only
+        ('$' in vm && vm.$.isUnmounted === false)
+      )
+      if (mounted) {
         watcher(getterOnly(vm.$options.asyncComputed[key]).apply(vm))
       }
     }
